backend/requirements.txt
# Core Flask dependencies
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-CORS>=4.0.0
Flask-Migrate>=4.0.0

# Authentication
PyJWT>=2.8.0
bcrypt>=4.1.0

# Database
psycopg2-binary>=2.9.0

# Environment
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-cov>=4.0.0

# Utilities
Werkzeug>=3.0.0
